CR - Move schemas closer to pages
