+++ src/app/_schemas/FrontmatterSchema.ts
@@ -4,8 +4,6 @@ import { z } from 'zod'
 
 import { isValidMarkdownPath } from '@/utils/fileUtils'
 
-import { createFeaturedProjectsSchema } from '@/ecosystem-explorer/utils/createFeaturedProjectsSchema'
-
 const FrontmatterHeaderSchema = z.object({
   title: z.string(),
   description: z.string().or(z.array(z.string())),
@@ -33,13 +31,3 @@ export const BaseFrontmatterSchema = z.object({
 export const FeaturedPageFrontmatterSchema = BaseFrontmatterSchema.extend({
   featured_entry: MarkdownPathSchema,
 })
-
-export const HomePageFrontmatterSchema = createFeaturedProjectsSchema(
-  'featured_ecosystem_projects',
-  'No matching ecosystem projects found',
-)
-
-export const GrantsPageFrontmatterSchema = createFeaturedProjectsSchema(
-  'featured_grant_graduates',
-  'No matching grant graduates found',
-)
